back_end/requirements.txt
asgiref==3.7.2
cryptography==42.0.5
Django==3.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
mysqlclient==2.2.1
pillow==10.2.0
pycryptodome==3.20.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0 
sqlparse==0.4.4
tzdata==2024.2